Major cleanup and refactor prior to PR
