+++ implementations/c/include/ockam/handshake.h
@@ -26,13 +26,7 @@
  */
 
 #define KEY_SIZE 32
-#define NAME_SIZE 28
 #define SHA256_SIZE 32
-#define NAME "Noise_XX_25519_AESGCM_SHA256"
-#define MAX_TRANSMIT_SIZE 2048
-#define DHLEN 32
-#define TAG_SIZE 16
-#define VECTOR_SIZE 12
 
 /*
  ********************************************************************************************************
@@ -41,7 +35,7 @@
  */
 
 /**
- * HANDSHAKE - the handshake structure is passed to all handshake functions.
+ * XX_HANDSHAKE - the handshake structure is passed to all handshake functions.
  */
 typedef struct  {
 
@@ -51,13 +45,13 @@ typedef struct  {
 	uint8_t     e[KEY_SIZE];
 	uint8_t     re[KEY_SIZE];
 	uint8_t     k[KEY_SIZE];
-	uint8_t     ck[SHA256_SIZE];
+	uint8_t     ck[KEY_SIZE];
 	uint8_t     h[SHA256_SIZE];
 	uint8_t     ke[KEY_SIZE];
 	uint8_t     kd[KEY_SIZE];
 	uint64_t    ne;
 	uint64_t    nd;
-} HANDSHAKE;
+} XX_HANDSHAKE;
 
 /*
  ********************************************************************************************************
@@ -70,26 +64,26 @@ typedef struct  {
  *                                      ockam_responder_handshake
  *
  * @param connection [in] - Initialize OCKAM_CONNECTION instance (must be connected)
- * @param p_h [in/out] - pointer to the HANDSHAKE structure. Should be 0-initialized prior to calling,
+ * @param p_h [in/out] - pointer to the XX_HANDSHAKE structure. Should be 0-initialized prior to calling,
  *                      and not modified thereafter.
  * @return [out] - OCKAM_ERR_NONE on success
  *
  * ********************************************************************************************************
  */
-OCKAM_ERR ockam_responder_handshake( OCKAM_TRANSPORT_CONNECTION connection, HANDSHAKE* p_h );
+OCKAM_ERR ockam_xx_responder_handshake( OCKAM_TRANSPORT_CONNECTION connection, XX_HANDSHAKE* p_h );
 
 /**
  * ********************************************************************************************************
  *                                      ockam_initiator_handshake
  *
  * @param connection [in] - Initialize OCKAM_CONNECTION instance (must be connected)
- * @param p_h [in/out] - pointer to the HANDSHAKE structure. Should be 0-initialized prior to calling,
+ * @param p_h [in/out] - pointer to the XX_HANDSHAKE structure. Should be 0-initialized prior to calling,
  *                      and not modified thereafter.
  * @return [out] - OCKAM_ERR_NONE on success
  *
  * ********************************************************************************************************
 */
-OCKAM_ERR ockam_initiator_handshake( OCKAM_TRANSPORT_CONNECTION connection, HANDSHAKE* p_h );
+OCKAM_ERR ockam_xx_initiator_handshake( OCKAM_TRANSPORT_CONNECTION connection, XX_HANDSHAKE* p_h );
 
 /**
  * ********************************************************************************************************
@@ -105,7 +99,7 @@ OCKAM_ERR ockam_initiator_handshake( OCKAM_TRANSPORT_CONNECTION connection, HAND
  *
  * ********************************************************************************************************
  */
-OCKAM_ERR decrypt( HANDSHAKE* p_h,
+OCKAM_ERR decrypt( XX_HANDSHAKE* p_h,
 		uint8_t* p_payload, uint32_t payload_size, uint8_t* p_msg, uint16_t msg_length, uint32_t* p_payload_bytes );
 
 /**
@@ -123,21 +117,8 @@ OCKAM_ERR decrypt( HANDSHAKE* p_h,
  *
  * ********************************************************************************************************
  */
- OCKAM_ERR encrypt( HANDSHAKE* p_h, uint8_t* p_payload, uint32_t payload_size,
+ OCKAM_ERR encrypt( XX_HANDSHAKE* p_h, uint8_t* p_payload, uint32_t payload_size,
                    uint8_t* p_msg, uint16_t msg_length, uint16_t* p_msg_size );
 
-void print_uint8_str( uint8_t* p, uint16_t size, char* msg );
-void string_to_hex(char* hexstring, uint8_t* val, uint32_t* p_bytes );
 
-OCKAM_ERR responder_m1_process( HANDSHAKE* p_h, uint8_t* p_m1, uint16_t m1_size );
-OCKAM_ERR responder_m2_make( HANDSHAKE* p_h, uint8_t* p_payload, uint32_t payload_size,
-                             uint8_t* p_msg, uint16_t msg_size, uint16_t* p_bytes_written );
-OCKAM_ERR responder_m3_process( HANDSHAKE* p_h, uint8_t* p_m3, uint16_t m3_size );
-OCKAM_ERR responder_epilogue( HANDSHAKE* p_h );
-OCKAM_ERR initiator_m1_make( HANDSHAKE* p_h, uint8_t* p_prologue, uint16_t prologue_length,
-                             uint8_t* p_payload, uint16_t payload_length, uint8_t* p_send_buffer, uint16_t buffer_length,
-                             uint16_t* p_transmit_size );
-OCKAM_ERR initiator_m2_process( HANDSHAKE* p_h, uint8_t* p_recv, uint16_t recv_size );
-OCKAM_ERR initiator_m3_make( HANDSHAKE* p_h, uint8_t* p_msg, uint16_t* p_msg_size );
-OCKAM_ERR initiator_epilogue( HANDSHAKE* p_h );
 #endif
