+++ implementations/c/source/ockam/handshake/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.13)
 project(ockam_handshake C)
 
-# Ockam Transport
+# Ockam Handshake
 
 ###########################
 # Set directory locations #
@@ -10,6 +10,8 @@ project(ockam_handshake C)
 set(COMMON_OCKAM_SRC_DIR $ENV{OCKAM_C_BASE}/source/ockam)
 set(OCKAM_HANDSHAKE_SRC_DIR ${CMAKE_CURRENT_SOURCE_DIR})
 set(OCKAM_INC_DIR ${OCKAM_C_BASE}/include)
+set(HANDSHAKE_INC_DIR ${OCKAM_INC_DIR}
+	${OCKAM_HANDSHAKE_SRC_DIR}/include)
 
 #################
 # Build Options #
@@ -21,8 +23,10 @@ set(OCKAM_INC_DIR ${OCKAM_C_BASE}/include)
 
 # Add the source files
 add_library(ockam_handshake STATIC
-  ${OCKAM_HANDSHAKE_SRC_DIR}/handshake.c
-  ${COMMON_OCKAM_SRC_DIR}/log/syslog.c)
+		${OCKAM_HANDSHAKE_SRC_DIR}/xx/xx_handshake.c
+		${OCKAM_HANDSHAKE_SRC_DIR}/xx/xx_initiator.c
+		${OCKAM_HANDSHAKE_SRC_DIR}/xx/xx_responder.c
+       ${COMMON_OCKAM_SRC_DIR}/log/syslog.c)
 
 # Set the include directories
